Update logic of price, product quantity, detailed cart quantity, when adding items to cart. As well as update email notification function using Emailsender when successfully creating an account.

+++ API_web_shop/src/main/java/web_shop/api_web_shop/entity/GioHangChiTiet.java
@@ -15,7 +15,7 @@ public class GioHangChiTiet {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long soLuong;
-    private Long thanhTien;
+    private Double thanhTien;
 
     @ManyToOne
     @JoinColumn(name = "san_pham_chi_tiet_id")
